shop/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
Django==5.0.4
django-celery-beat==2.6.0
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-filter==24.2
django-humanize==0.1.2
django-jalali==6.2.0
django-jalali-date==1.1.3
django-jazzmin==2.6.1
django-js-asset==2.2.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.1.1
hiredis==3.0.0
humanize==4.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.2
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.31.0
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
pycodestyle==2.12.1
pylint==3.2.7